+++ R/construct.r
@@ -6,12 +6,11 @@
 #' @param data           A data frame
 #' @param formula        A formula: value ~ axis1 [+ axis2 + axis n ..]
 #' @param fill           Value to fill array with if undefined
-#' @param fun.aggregate  Function to aggregate multiple values for the same position; default: error
-#' @param ...            Additional arguments passed to reshape2::acast
+#' @param name_axes      Keep column names of `data` as axis names
 #' @return               A structured array
 #' @export
-construct = function(data, formula=guess_structure(data), fill=NULL,
-                     fun.aggregate=NULL, ...) {
+construct = function(data, formula=guess_structure(data), fill=NA,
+                     name_axes=TRUE) {
     if (!is.data.frame(data))
         stop("`data` needs to reference a data.frame")
 
@@ -20,19 +19,25 @@ construct = function(data, formula=guess_structure(data), fill=NULL,
         stop("The dependent variable (left side of ~) needs to reference exactly one variable")
 
     indep_vars = all.vars(formula[[3]]) #TODO: if factor, include all levels in matrix
-    form = stats::as.formula(paste(indep_vars, collapse = "~"))
 
-    data = as.data.frame(data)
     axis_NA = apply(is.na(data[indep_vars]), 1, any)
     if (any(axis_NA)) {
         warning("Omitting ", sum(axis_NA), " rows where axis columns are NA")
         data = data[!axis_NA,]
     }
 
-    reshape2::acast(data,
-                    formula = form,
-                    value.var = dep_var,
-                    fill = fill,
-                    fun.aggregate = fun.aggregate,
-                    ...) %catchm% stop("Do not know how to aggregate")
+    axes = data[indep_vars]
+    values = data[[dep_var]]
+
+    if (any(duplicated(axes)))
+        stop("Duplicated entries in `data` are not allowed")
+
+    dimNames = lapply(axes, unique)
+    ndim = c(unname(sapply(dimNames, length)))
+    result = array(fill, dim=ndim, dimnames=dimNames)
+
+    for (i in seq_len(nrow(data)))
+        result = do.call("[<-", c(list(result), axes[i,], list(values[[i]])))
+
+    result
 }