+++ CHANGES.md
@@ -1,9 +1,12 @@
 * 0.4.0
   * `stack` now also accepts multiple arrays as arguments
   * `summarize` is kept as `translate`
-  * `NA` subsets in `map` are dropped with a warning (#13)
+  * `construct` no longer allows value aggregation; do this on the `data.frame`
+    before (related to #11)
+  * `construct` now as `name_axes` option to keep column names (#12)
+  * `map` subsets that are `NA` are dropped with a warning (#13)
   * `map` subsets throw error if not same length as array axis
-  * new lambda syntax (#14)
+  * new `lambda` syntax (#14)
 
 * 0.3.2
   * add `collect` function as opposite to `mask`